+++ document/release-notes.md
@@ -1,3 +1,6 @@
+### 3.3.0.79
+> Automatically execute pragma on connection openning
+> Clone will create new connection
+
 ### 3.2.1.78
 > The ```LiteralValue<T> where T : class``` added.
-